requirements.txt
# Core Dependencies
python-dotenv==1.0.0
psycopg2-binary==2.9.9

# LangChain & AI
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1

# Data & Finance
yfinance==0.2.35
pandas==2.1.4
numpy==1.26.2
pandas-ta==0.3.14b0

# API & HTTP
requests==2.31.0

# Machine Learning
scikit-learn>=1.3.2
joblib>=1.3.0
joblib==1.3.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
